requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
cramjam==2.5.0
discord.py==2.0.1
exceptiongroup==1.0.0rc9
fastparquet==0.8.3
frozenlist==1.3.1
fsspec==2022.8.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
idna==3.4
lxml==4.9.1
multidict==6.0.2
nba-api==1.1.11
nfl-data-py==0.3.0
nflgame==1.2.20
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2022.2.1
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.1
